Move v1 storage adapters to a nested dedicated directory
