+++ src/globus_sdk/tokenstorage/__init__.py
@@ -1,12 +1,16 @@
-from globus_sdk.tokenstorage.base import FileAdapter, StorageAdapter
-from globus_sdk.tokenstorage.file_adapters import SimpleJSONFileAdapter
-from globus_sdk.tokenstorage.memory_adapter import MemoryAdapter
-from globus_sdk.tokenstorage.sqlite_adapter import SQLiteAdapter
+from .v1 import (
+    FileAdapter,
+    MemoryAdapter,
+    SimpleJSONFileAdapter,
+    SQLiteAdapter,
+    StorageAdapter,
+)
 
 __all__ = (
-    "SimpleJSONFileAdapter",
-    "SQLiteAdapter",
+    # v1 "StorageAdapter" Constructs
     "StorageAdapter",
     "FileAdapter",
+    "SimpleJSONFileAdapter",
+    "SQLiteAdapter",
     "MemoryAdapter",
 )
