+++ src/globus_sdk/tokenstorage/v1/file_adapters.py
@@ -5,7 +5,7 @@ import pathlib
 import typing as t
 
 from globus_sdk.services.auth import OAuthTokenResponse
-from globus_sdk.tokenstorage.base import FileAdapter
+from globus_sdk.tokenstorage.v1.base import FileAdapter
 from globus_sdk.version import __version__
 
 # use the non-annotation form of TypedDict to apply a non-identifier key
